app-engine/GCP-App-Engine-standard/python-bigquery-simple/requirements.txt
google-cloud-bigquery==2.34.4
Flask==2.0.3
Werkzeug==2.0.3
gunicorn==20.1.0